+++ docs/versionhistory.rst
@@ -14,6 +14,8 @@ This library adheres to
 - Fixed ``typing`` types (``dict[str, int]``, ``List[str]``, etc.) not passing checks
   against ``type`` or ``Type``
   (`#432 <https://github.com/agronholm/typeguard/issues/432>`_, PR by Yongxin Wang)
+- Fixed detection of optional fields (``NotRequired[...]``) in ``TypedDict`` when using
+  forward references (`#424 <https://github.com/agronholm/typeguard/issues/424>`_)
 
 **4.1.5** (2023-09-11)